// frontend/src/api/models/req/update-acl-entry-request.ts
/**
 * Generated by orval v8.4.2 🍺
 * Do not edit manually.
 * Ello Bot
 * Ello Bot Backend API
 * OpenAPI spec version: 1.0.0
 */

export interface UpdateAclEntryRequest {
  permissionId?: number | null;
  subjectPrincipalId?: number | null;
  subjectRoleId?: number | null;
  /**
   * @minLength 1
   * @maxLength 16
   */
  effect?: string | null;
}