// frontend/src/api/models/resp/acl-entry-response.ts
/**
 * Generated by orval v8.4.2 🍺
 * Do not edit manually.
 * Ello Bot
 * Ello Bot Backend API
 * OpenAPI spec version: 1.0.0
 */

export interface AclEntryResponse {
  id: number;
  tenantId: number;
  resourceType: string;
  resourceId: number;
  permissionId: number;
  subjectPrincipalId?: number | null;
  subjectRoleId?: number | null;
  effect: string;
  createdAt: string;
  updatedAt: string;
}