// frontend/src/api/models/resp/role-response.ts
/**
 * Generated by orval v8.4.2 🍺
 * Do not edit manually.
 * Ello Bot
 * Ello Bot Backend API
 * OpenAPI spec version: 1.0.0
 */
import type { PermissionResponse } from './permission-response';

export interface RoleResponse {
  id: number;
  tenantId: number;
  code: string;
  name: string;
  description: string;
  isBuiltin: boolean;
  createdAt: string;
  updatedAt: string;
  permissions?: PermissionResponse[];
}